+++ src/Bibliotheca.Server.Gateway.Api/Controllers/ProjectsController.cs
@@ -1,5 +1,6 @@
 ﻿using System.Threading.Tasks;
 using Bibliotheca.Server.Gateway.Core.DataTransferObjects;
+using Bibliotheca.Server.Gateway.Core.Policies;
 using Bibliotheca.Server.Gateway.Core.Services;
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
@@ -13,9 +14,12 @@ namespace Bibliotheca.Server.Gateway.Api.Controllers
     {
         private readonly IProjectsService _projectsService;
 
-        public ProjectsController(IProjectsService projectsService)
+        private readonly IAuthorizationService _authorizationService;
+
+        public ProjectsController(IProjectsService projectsService, IAuthorizationService authorizationService)
         {
             _projectsService = projectsService;
+            _authorizationService = authorizationService;
         }
 
         [HttpGet()]
@@ -26,13 +30,19 @@ namespace Bibliotheca.Server.Gateway.Api.Controllers
         }
 
         [HttpGet("{projectId}")]
-        public async Task<ProjectDto> Get(string projectId)
+        public async Task<IActionResult> Get(string projectId)
         {
             var project = await _projectsService.GetProjectAsync(projectId);
-            return project;
+            if (project == null)
+            {
+                return NotFound();
+            }
+
+            return new ObjectResult(project);
         }
 
         [HttpPost]
+        [Authorize("CanAddProject")]
         public async Task<IActionResult> Post([FromBody] ProjectDto project)
         {
             await _projectsService.CreateProjectAsync(project);
@@ -42,6 +52,12 @@ namespace Bibliotheca.Server.Gateway.Api.Controllers
         [HttpPut("{projectId}")]
         public async Task<IActionResult> Put(string projectId, [FromBody] ProjectDto project)
         {
+            var isAuthorize = await _authorizationService.AuthorizeAsync(User, new ProjectDto { Id = projectId }, Operations.Update);
+            if (!isAuthorize)
+            {
+                return Forbid();
+            }
+
             await _projectsService.UpdateProjectAsync(projectId, project);
             return Ok();
         }
@@ -49,6 +65,12 @@ namespace Bibliotheca.Server.Gateway.Api.Controllers
         [HttpDelete("{projectId}")]
         public async Task<IActionResult> Delete(string projectId)
         {
+            var isAuthorize = await _authorizationService.AuthorizeAsync(User, new ProjectDto { Id = projectId }, Operations.Delete);
+            if (!isAuthorize)
+            {
+                return Forbid();
+            }
+
             await _projectsService.DeleteProjectAsync(projectId);
             return Ok();
         }
